fix: keep conductor storage under static safety limit
